Use float32

+++ finemo/hitcaller.py
@@ -163,10 +163,15 @@ def fit_contribs(cwms, contribs, sequences,
         end = min(n, start + batch_size)
         b = end - start
 
+        # sequences_batch = F.pad(sequences[start:end,:,:], (0, w - 1)).to(device=device) # (b, 4, l + w - 1)
+        # contribs_batch = F.pad(contribs[start:end,None,:], (0, w - 1)).to(device=device) 
+        # contribs_batch = (contribs_batch / contrib_norm) * sequences_batch.half() # (b, 4, l + w - 1)
+        # coef_init = torch.zeros((b, m, l + 2 * w - 2), dtype=torch.float16, device=device) # (b, m, l + 2w - 2)
+
         sequences_batch = F.pad(sequences[start:end,:,:], (0, w - 1)).to(device=device) # (b, 4, l + w - 1)
-        contribs_batch = F.pad(contribs[start:end,None,:], (0, w - 1)).to(device=device) 
-        contribs_batch = (contribs_batch / contrib_norm) * sequences_batch.half() # (b, 4, l + w - 1)
-        coef_init = torch.zeros((b, m, l + 2 * w - 2), dtype=torch.float16, device=device) # (b, m, l + 2w - 2)
+        contribs_batch = F.pad(contribs[start:end,None,:], (0, w - 1)).float().to(device=device) 
+        contribs_batch = (contribs_batch / contrib_norm) * sequences_batch # (b, 4, l + w - 1)
+        coef_init = torch.zeros((b, m, l + 2 * w - 2), dtype=torch.float32, device=device) # (b, m, l + 2w - 2)
 
         coef, ll, gap, steps = fit_batch(cwms, cwms_t, contribs_batch, sequences_batch, coef_init, clip_mask,
                                          a_const, b_const, step_size, convergence_tol, max_steps)
